+++ app/Http/routes.php
@@ -87,6 +87,7 @@ if (config('dfe.enable-fast-track')) {
 /** Login event listener */
 \Event::listen('auth.login',
     function() {
+        /** @noinspection PhpUndefinedMethodInspection */
         \Auth::user()->update([
             'last_login_date'    => date('c'),
             'last_login_ip_text' => \Request::server('REMOTE_ADDR'),